requirements.txt
flask 
pymongo
pycryptodome>=3.15.0